Refactor and consolidate unit tests for venue normalization and author handling
- Removed outdated test files for proceedings ordinal normalization, Semantic Scholar API URL regression, validation bug fixes, and venue parsing regression.
- Integrated tests for proceedings normalization with ordinal numbers into a new consolidated test class.
- Added comprehensive tests for handling "and others" in author lists, ensuring correct parsing and edge case handling.
- Enhanced author comparison tests to address deduplication and error message accuracy.
- Improved venue parsing tests to handle LaTeX commands and volume/page information normalization.
- Ensured backwards compatibility for existing functionality while introducing new regression tests.
