+++ src/core/refchecker.py
@@ -1932,23 +1932,27 @@ class ArxivReferenceChecker:
             })
         
         # Verify DOI
-        if doi and external_ids.get('DOI') and doi.lower() != external_ids['DOI'].lower():
-            # Check if the cited DOI is a partial match of the actual DOI
-            # This handles cases like "10.1111/j.2044-8260." vs "10.1111/J.2044-8260.1997.TB01237.X"
-            cited_doi_clean = doi.lower().rstrip('.')
-            actual_doi_clean = external_ids['DOI'].lower().rstrip('.')
+        if doi and external_ids.get('DOI'):
+            from utils.doi_utils import compare_dois, normalize_doi
             
-            # If the cited DOI is a prefix of the actual DOI, it's likely a partial citation
-            # Only flag as error if it's not a reasonable partial match
-            if not actual_doi_clean.startswith(cited_doi_clean):
-                logger.debug(f"DB Verification: DOI mismatch - cited: {doi}, actual: {external_ids['DOI']}")
-                errors.append({
-                    'error_type': 'doi',
-                    'error_details': f"DOI mismatch: cited as {doi} but actually {external_ids['DOI']}",
-                    'ref_doi_correct': external_ids['DOI']
-                })
-            else:
-                logger.debug(f"DB Verification: DOI partial match - cited: {doi}, actual: {external_ids['DOI']} (acceptable)")
+            # Use proper DOI comparison first
+            if not compare_dois(doi, external_ids['DOI']):
+                # Check if the cited DOI is a partial match of the actual DOI
+                # This handles cases like "10.1111/j.2044-8260." vs "10.1111/J.2044-8260.1997.TB01237.X"
+                cited_doi_normalized = normalize_doi(doi)
+                actual_doi_normalized = normalize_doi(external_ids['DOI'])
+                
+                # If the cited DOI is a prefix of the actual DOI, it's likely a partial citation
+                # Only flag as error if it's not a reasonable partial match
+                if not actual_doi_normalized.startswith(cited_doi_normalized.rstrip('.')):
+                    logger.debug(f"DB Verification: DOI mismatch - cited: {doi}, actual: {external_ids['DOI']}")
+                    errors.append({
+                        'error_type': 'doi',
+                        'error_details': f"DOI mismatch: cited as {doi} but actually {external_ids['DOI']}",
+                        'ref_doi_correct': external_ids['DOI']
+                    })
+                else:
+                    logger.debug(f"DB Verification: DOI partial match - cited: {doi}, actual: {external_ids['DOI']} (acceptable)")
 
         # Verify ArXiv ID
         if reference.get('type') == 'arxiv':
